backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "local-music-player"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "celery[redis]>=5.4.0",
    "redis>=5.0.4",
    "yt-dlp>=2024.5.1",
    "pydantic-settings>=2.2.1",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
